made all frontend changes
