+++ src/container/App.jsx
@@ -5,20 +5,20 @@ import { BrowserRouter as Router, Route, Switch } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import axios from 'axios';
 import { connect } from 'react-redux';
-import API_URL from './helpers/apiUrl';
-import Menu from './Menu';
-import MobileMenu from './MobileMenu';
-import Home from './Home';
-import Providers from './provider/Providers';
-import ProviderComponent from './provider/ProviderComponent';
-import NewProvider from './provider/NewProvider';
+import API_URL from '../components/helpers/apiUrl';
+import Menu from '../components/Menu';
+import MobileMenu from '../components/MobileMenu';
+import Home from '../components/Home';
+import Providers from './providers/Providers';
+import ProviderComponent from './providers/ProviderComponent';
+import NewProvider from './providers/NewProvider';
 import NewAppointment from './appointments/NewAppointment';
-import EditProvider from './provider/EditProvider';
+import EditProvider from './providers/EditProvider';
 import Registration from './auth/Registration';
 import Login from './auth/Login';
 import Appointments from './appointments/Appointments';
 import UserAppointment from './appointments/UserAppointment';
-import './styles/App.css';
+import '../components/styles/App.css';
 import { storeUser } from '../actions/index';
 
 class App extends Component {
