+++ src/index.js
@@ -5,7 +5,7 @@ import { createStore } from 'redux';
 import { Provider } from 'react-redux';
 import rootReducer from './reducers';
 
-import App from './components/App';
+import App from './container/App';
 import './components/styles/reset.css';
 
 const store = createStore(rootReducer, { user: {} });
